feat(analytics): move AnalyticsRepository to Persistence layer, rewire DI
