+++ backend/test/Anela.Heblo.Tests/Features/Analytics/AnalyticsRepositoryTests.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
-using Anela.Heblo.Application.Features.Analytics.Infrastructure;
+using Anela.Heblo.Persistence.Features.Analytics;
 using Anela.Heblo.Domain.Features.Analytics;
 using FluentAssertions;
 using Moq;
